pyproject.toml
[tool.poetry]
name = "bord"
version = "0.1.0"
description = ""
authors = ["Steinn Eldjárn Sigurðarson <steinnes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
starlette = "^0.13.1"
pyppeteer2 = "=0.2.2"
urllib3 = "^1.25"
screeninfo = "^0.6.1"
asyncclick = "^7.0"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}

[tool.poetry.scripts]
bord = 'bord.cli:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
